+++ web/apps/photos/src/services/export/migration.ts
@@ -2,13 +2,12 @@ import { ensureElectron } from "@/base/electron";
 import { nameAndExtension } from "@/base/file";
 import log from "@/base/log";
 import type { Collection } from "@/media/collection";
-import { EnteFile } from "@/media/file";
+import { type EnteFile, mergeMetadata } from "@/media/file";
 import { FileType } from "@/media/file-type";
 import { decodeLivePhoto } from "@/media/live-photo";
 import downloadManager from "@/new/photos/services/download";
 import { exportMetadataDirectoryName } from "@/new/photos/services/export";
 import { getAllLocalFiles } from "@/new/photos/services/files";
-import { mergeMetadata } from "@/new/photos/utils/file";
 import {
     safeDirectoryName,
     safeFileName,
